fix: kill duplicate processes

+++ megabot.py
@@ -282,8 +282,9 @@ if __name__ == "__main__":
     async def call(ctx):
         await randomemoji(ctx)
 
-    pids = list(map(int, check_output(["pidof", "python3"]).split())).remove(str(os.getpid()))
+    pids = list(map(int, check_output(["pidof", "python3"]).split()))
     for p in pids:
-        os.kill(p)
+        if str(os.getpid()) != p:
+            os.kill(p)
 
     bot.run(TOKEN)
